chore(#043): Change register tests from "test.py" to "tests.py" and delete "test.py"
